Add endpoint for json editor
Add save endpoint
Rearrange table/save file to table folder
