+++ frictionless/project/project.py
@@ -139,6 +139,9 @@ class Project:
         self.database.move_record(source, target)
         return target
 
+    def save_file(self, name: str, *, bytes: bytes, folder: Optional[str] = None) -> str:
+        return self.filesystem.create_file(name, bytes=bytes, folder=folder)
+
     # TODO: implement
     def update_file(self, path: str):
         pass
