+++ frictionless/server/file/save.py
@@ -1,22 +1,28 @@
 from typing import Optional
 from pydantic import BaseModel
-from fastapi import Request
+from fastapi import Request, UploadFile, File, Form
 from ...project import Project
 from ..router import router
 
 
+# See the signature
 class Props(BaseModel):
-    session: Optional[str]
-    path: str
-    tablePatch: dict
+    pass
 
 
 class Result(BaseModel):
     path: str
 
 
-@router.post("/table/save")
-def server_table_save(request: Request, props: Props) -> Result:
-    project: Project = request.app.get_project(props.session)
-    path = project.save_table(props.path, tablePatch=props.tablePatch)
+@router.post("/file/save")
+async def server_file_save(
+    request: Request,
+    file: UploadFile = File(),
+    folder: Optional[str] = Form(None),
+    session: Optional[str] = Form(None),
+) -> Result:
+    project: Project = request.app.get_project(session)
+    name = file.filename or "name"
+    bytes = await file.read()
+    path = project.save_file(name, bytes=bytes, folder=folder)
     return Result(path=path)
