+++ frictionless/server/table/__init__.py
@@ -2,3 +2,4 @@
 from . import export
 from . import query
 from . import read
+from . import save
